requirements.txt
# Runtime
flask>=3.0.0,<4
gunicorn>=22.0.0,<24
pulp>=2.8.0,<3
python-dotenv>=1.0.0,<2

# Dev / test
pytest>=8.0.0,<9
pytest-cov>=5.0.0,<6
